requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.0
python-multipart==0.0.12
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
